alz-thinktank/requirements.txt
pandas
numpy
lifelines
scikit-learn
matplotlib
pyyaml
pyarrow
